refactor: extract query pagination and output formatting helpers
