+++ src/cscout.cpp
@@ -87,6 +87,8 @@
 #include "timer.h"
 #include "dbtoken.h"
 #include "macro_arg_processor.h"
+#include "query_pagination.h"
+#include "query_text_output.h"
 
 #ifdef PICO_QL
 #include "pico_ql_search.h"
@@ -1014,117 +1016,6 @@ html_file(FILE *of, Fileid fi)
 		fname.c_str());
 }
 
-struct Pagination_data {
-	int skip;
-	int pagesize;
-	bool show_all;
-};
-
-static bool
-json_output_requested()
-{
-	return !!swill_getvar("json");
-}
-
-static Pagination_data
-pagination_data()
-{
-	Pagination_data p;
-	if (!swill_getargs("I(skip)", &p.skip))
-		p.skip = 0;
-	p.pagesize = Option::entries_per_page->get();
-	p.show_all = (p.skip == -1);
-	return p;
-}
-
-static bool
-in_page(const Pagination_data &p, int index)
-{
-	return p.show_all || (index >= p.skip && index < p.skip + p.pagesize);
-}
-
-static string
-json_page_url(const string &base_url, int skip)
-{
-	ostringstream s;
-	s << base_url;
-	if (base_url.find("json=1") == string::npos)
-		s << "&json=1";
-	s << "&skip=" << skip;
-	return s.str();
-}
-
-static void
-json_puts(FILE *of, const string &value)
-{
-	fputc('"', of);
-	for (string::const_iterator i = value.begin(); i != value.end(); i++) {
-		unsigned char c = static_cast<unsigned char>(*i);
-		switch (c) {
-		case '"':
-			fputs("\\\"", of);
-			break;
-		case '\\':
-			fputs("\\\\\\\\", of);
-			break;
-		case '\b':
-			fputs("\\\\b", of);
-			break;
-		case '\f':
-			fputs("\\\\f", of);
-			break;
-		case '\n':
-			fputs("\\\\n", of);
-			break;
-		case '\r':
-			fputs("\\\\r", of);
-			break;
-		case '\t':
-			fputs("\\\\t", of);
-			break;
-		default:
-			if (c < 0x20)
-				fprintf(of, "\\\\u%04x", c);
-			else
-				fputc(c, of);
-		}
-	}
-	fputc('"', of);
-}
-
-static void
-json_output_pagination(FILE *of, const string &base_url, const Pagination_data &p, int total, int returned)
-{
-	int effective_skip = p.show_all ? 0 : p.skip;
-	int npages = total / p.pagesize + (total % p.pagesize ? 1 : 0);
-	int current_page = p.show_all ? 1 : effective_skip / p.pagesize + 1;
-	bool has_prev = !p.show_all && effective_skip > 0;
-	bool has_next = !p.show_all && effective_skip + p.pagesize < total;
-
-	fputs("\"pagination\":{", of);
-	fprintf(of, "\"total\":%d,\"returned\":%d,\"skip\":%d,\"page_size\":%d,\"show_all\":%s,\"page\":%d,\"pages\":%d,\"has_prev\":%s,\"has_next\":%s",
-		total,
-		returned,
-		effective_skip,
-		p.pagesize,
-		p.show_all ? "true" : "false",
-		npages ? current_page : 0,
-		npages,
-		has_prev ? "true" : "false",
-		has_next ? "true" : "false");
-	if (has_prev) {
-		fputs(",\"prev\":", of);
-		json_puts(of, json_page_url(base_url, effective_skip - p.pagesize));
-	}
-	if (has_next) {
-		fputs(",\"next\":", of);
-		json_puts(of, json_page_url(base_url, effective_skip + p.pagesize));
-	}
-	fputs(",\"all\":", of);
-	json_puts(of, json_page_url(base_url, -1));
-	fputc('}', of);
-}
-
 // File query page
 static int
 filequery_page(FILE *of,  void *)
@@ -1182,7 +1073,7 @@ xfilequery_page(FILE *of,  void *)
 		int returned = 0;
 
 		fputs("{", of);
-		for (multiset <Fileid, FileQuery::specified_order>::iterator i = sorted_files.begin(); i != sorted_files.end(); i++) {
+		for (multiset <Fileid, FileQuery::FileComparator>::iterator i = sorted_files.begin(); i != sorted_files.end(); i++) {
 			Fileid f = *i;
 			if (current_project && !Filedetails::get_attribute(f, current_project))
 				continue;
@@ -1203,7 +1094,7 @@ xfilequery_page(FILE *of,  void *)
 		fputc('}', of);
 	} else if (plain_text) {
 		Pager pager(of, INT_MAX, query.base_url(), query.bookmarkable());
-		for (multiset <Fileid, FileQuery::specified_order>::iterator i = sorted_files.begin(); i != sorted_files.end(); i++) {
+		for (multiset <Fileid, FileQuery::FileComparator>::iterator i = sorted_files.begin(); i != sorted_files.end(); i++) {
 			Fileid f = *i;
 			if (current_project && !Filedetails::get_attribute(f, current_project))
 				continue;
@@ -1222,7 +1113,7 @@ xfilequery_page(FILE *of,  void *)
 
 		Pager pager(of, Option::entries_per_page->get(), query.base_url(), query.bookmarkable());
 		html_file_set_begin(of);
-		for (multiset <Fileid, FileQuery::specified_order>::iterator i = sorted_files.begin(); i != sorted_files.end(); i++) {
+		for (multiset <Fileid, FileQuery::FileComparator>::iterator i = sorted_files.begin(); i != sorted_files.end(); i++) {
 			Fileid f = *i;
 			if (current_project && !Filedetails::get_attribute(f, current_project))
 				continue;
@@ -1244,114 +1135,6 @@ xfilequery_page(FILE *of,  void *)
 	return 0;
 }
 
-
-/*
- * Display the sorted identifiers or functions, taking into account the reverse sort property
- * for properly aligning the output.
- */
-template <typename container>
-static void
-display_sorted(FILE *of, const Query &query, const container &sorted_ids)
-{
-	bool plain_text = !!swill_getvar("txt");
-	bool json_output = json_output_requested();
-
-	if (json_output) {
-		Pagination_data p = pagination_data();
-		int total = 0;
-		int returned = 0;
-
-		fputs("{\"results\":[", of);
-		for (typename container::const_iterator i = sorted_ids.begin(); i != sorted_ids.end(); i++) {
-			if (in_page(p, total)) {
-				if (returned++)
-					fputc(',', of);
-				output_json_element(of, *i);
-			}
-			total++;
-		}
-		fputs("],", of);
-		json_output_pagination(of, query.base_url() + "&qi=1", p, total, returned);
-		fputc('}', of);
-		return;
-	}
-
-	if (plain_text) {
-		Pager pager(of, INT_MAX, query.base_url() + "&qi=1", query.bookmarkable());
-		typename container::const_iterator i;
-		for (i = sorted_ids.begin(); i != sorted_ids.end(); i++) {
-			if (pager.show_next()) {
-				output_plain_text_element(of, *i);
-				fputc('\n', of);
-			}
-		}
-		return;
-	} else {
-		if (Option::sort_rev->get())
-			fputs("<table><tr><td width=\"50%\" align=\"right\">\n", of);
-		else
-			fputs("<p>\n", of);
-
-		Pager pager(of, Option::entries_per_page->get(), query.base_url() + "&qi=1", query.bookmarkable());
-		typename container::const_iterator i;
-		for (i = sorted_ids.begin(); i != sorted_ids.end(); i++) {
-			if (pager.show_next()) {
-				html(of, **i);
-				fputs("<br>\n", of);
-			}
-		}
-
-		if (Option::sort_rev->get())
-			fputs("</td> <td width=\"50%\"> </td></tr></table>\n", of);
-		else
-			fputs("</p>\n", of);
-		pager.end();
-	}
-}
-
-// Helper function to output plain text for different element types
-template <typename T>
-void output_plain_text_element(FILE *of, const T &element) {
-	fputs(element.c_str(), of);
-}
-
-template <typename T>
-void output_json_element(FILE *of, const T &element) {
-	json_puts(of, element.c_str());
-}
-
-// Specialization for identifier pairs (Eclass*, Identifier)
-template <>
-void output_plain_text_element(FILE *of, const std::pair<Eclass* const, Identifier> &element) {
-	fputs(element.second.get_id().c_str(), of);
-}
-
-template <>
-void output_json_element(FILE *of, const std::pair<Eclass* const, Identifier> &element) {
-	json_puts(of, element.second.get_id());
-}
-
-// Specialization for function pointers (Call*)
-template <>
-void output_plain_text_element(FILE *of, const Call* const &element) {
-	fputs(element->get_name().c_str(), of);
-}
-
-template <>
-void output_json_element(FILE *of, const Call* const &element) {
-	json_puts(of, element->get_name());
-}
-
-// Specialization for pointer to identifier pairs (const std::pair<Eclass* const, Identifier>*)
-template <>
-void output_plain_text_element(FILE *of, const std::pair<Eclass* const, Identifier>* const &element) {
-	fputs(element->second.get_id().c_str(), of);
-}
-
-template <>
-void output_json_element(FILE *of, const std::pair<Eclass* const, Identifier>* const &element) {
-	json_puts(of, element->second.get_id());
-}
 
 /*
  * Display the sorted functions with their metrics,
